更新

+++ client/client.realize/messengers/clients/ClientsMessenger.cs
@@ -9,7 +9,7 @@ namespace client.realize.messengers.clients
     /// <summary>
     /// 服务器发来的客户端列表
     /// </summary>
-    [MessengerIdRange((int)ClientsMessengerIds.Min, (int)ClientsMessengerIds.Max)]
+    [MessengerIdRange((ushort)ClientsMessengerIds.Min, (ushort)ClientsMessengerIds.Max)]
     public class ClientsMessenger : IMessenger
     {
         private readonly ClientsMessengerSender clientsMessengerSender;
@@ -20,7 +20,7 @@ namespace client.realize.messengers.clients
             this.registerTransfer = registerTransfer;
         }
 
-        [MessengerId((int)ClientsMessengerIds.Notify)]
+        [MessengerId((ushort)ClientsMessengerIds.Notify)]
         public void Notify(IConnection connection)
         {
             ClientsInfo res = new ClientsInfo();
@@ -28,7 +28,7 @@ namespace client.realize.messengers.clients
             clientsMessengerSender.OnServerClientsData.Push(res);
         }
 
-        [MessengerId((int)ClientsMessengerIds.Reset)]
+        [MessengerId((ushort)ClientsMessengerIds.Reset)]
         public async Task Reset(IConnection connection)
         {
             await registerTransfer.Register().ConfigureAwait(false);
