+++ client/client.realize/messengers/punchHole/PunchHoleMessenger.cs
@@ -6,7 +6,7 @@ using System;
 
 namespace client.realize.messengers.punchHole
 {
-    [MessengerIdRange((int)PunchHoleMessengerIds.Min, (int)PunchHoleMessengerIds.Max)]
+    [MessengerIdRange((ushort)PunchHoleMessengerIds.Min, (ushort)PunchHoleMessengerIds.Max)]
     public class PunchHoleMessenger : IMessenger
     {
         private readonly PunchHoleMessengerSender punchHoleMessengerSender;
@@ -16,7 +16,7 @@ namespace client.realize.messengers.punchHole
             this.punchHoleMessengerSender = punchHoleMessengerSender;
         }
 
-        [MessengerId((int)PunchHoleMessengerIds.Execute)]
+        [MessengerId((ushort)PunchHoleMessengerIds.Execute)]
         public void Execute(IConnection connection)
         {
             PunchHoleParamsInfo model = new PunchHoleParamsInfo();
