+++ client/plugins/client.service.ui/client.service.ui.api.manager/CounterClientService.cs
@@ -27,7 +27,7 @@ namespace client.service.ui.api.manager
                 var resp = await messengerSender.SendReply(new MessageRequestWrap
                 {
                     Payload = Helper.EmptyArray,
-                    MessengerId = (int)CounterMessengerIds.Info,
+                    MessengerId = (ushort)CounterMessengerIds.Info,
                     Connection = registerState.OnlineConnection,
                     Timeout = 15000
                 }).ConfigureAwait(false);
