+++ client/plugins/client.service.vea/VeaSocks5Messenger.cs
@@ -4,7 +4,7 @@ using System;
 
 namespace client.service.vea
 {
-    [MessengerIdRange((int)VeaSocks5MessengerIds.Min,(int)VeaSocks5MessengerIds.Max)]
+    [MessengerIdRange((ushort)VeaSocks5MessengerIds.Min,(ushort)VeaSocks5MessengerIds.Max)]
     public class VeaSocks5Messenger : Socks5Messenger
     {
         public VeaSocks5Messenger(IVeaSocks5ClientHandler socks5ClientHandler, IVeaSocks5ServerHandler socks5ServerHandler)
@@ -13,13 +13,13 @@ namespace client.service.vea
 
         }
 
-        [MessengerId((int)VeaSocks5MessengerIds.Request)]
+        [MessengerId((ushort)VeaSocks5MessengerIds.Request)]
         public new void Request(IConnection connection)
         {
             base.Request(connection);
         }
 
-        [MessengerId((int)VeaSocks5MessengerIds.Response)]
+        [MessengerId((ushort)VeaSocks5MessengerIds.Response)]
         public new void Response(IConnection connection)
         {
             base.Response(connection);
@@ -27,7 +27,7 @@ namespace client.service.vea
     }
 
     [Flags, MessengerIdEnum]
-    public enum VeaSocks5MessengerIds : int
+    public enum VeaSocks5MessengerIds : ushort
     {
         Min = 901,
         Request = 902,
