+++ client/plugins/client.service.vea/VeaSocks5MessengerSender.cs
@@ -11,8 +11,8 @@ namespace client.service.vea
 
     public class VeaSocks5MessengerSender : Socks5MessengerSender, IVeaSocks5MessengerSender
     {
-        protected override int TargetRequest { get; } = (int)VeaSocks5MessengerIds.Request;
-        protected override int TargetResponse { get; } = (int)VeaSocks5MessengerIds.Response;
+        protected override ushort TargetRequest { get; } = (ushort)VeaSocks5MessengerIds.Request;
+        protected override ushort TargetResponse { get; } = (ushort)VeaSocks5MessengerIds.Response;
         public VeaSocks5MessengerSender(MessengerSender messengerSender) : base(messengerSender)
         {
         }
