+++ common/common.socks5/Socks5Messenger.cs
@@ -2,7 +2,7 @@
 
 namespace common.socks5
 {
-    [MessengerIdRange((int)Socks5MessengerIds.Min, (int)Socks5MessengerIds.Max)]
+    [MessengerIdRange((ushort)Socks5MessengerIds.Min, (ushort)Socks5MessengerIds.Max)]
     public class Socks5Messenger : IMessenger
     {
         private readonly ISocks5ClientHandler socks5ClientHandler;
@@ -13,13 +13,13 @@ namespace common.socks5
             this.socks5ServerHandler = socks5ServerHandler;
         }
 
-        [MessengerId((int)Socks5MessengerIds.Request)]
+        [MessengerId((ushort)Socks5MessengerIds.Request)]
         public void Request(IConnection connection)
         {
             socks5ServerHandler.InputData(connection);
         }
 
-        [MessengerId((int)Socks5MessengerIds.Response)]
+        [MessengerId((ushort)Socks5MessengerIds.Response)]
         public void Response(IConnection connection)
         {
             socks5ClientHandler.InputData(connection);
